add book controller, book view

+++ src/main/java/com/nashtech/intern/dao/IBookDAO.java
@@ -6,8 +6,8 @@ import com.nashtech.intern.model.Book;
 
 public interface IBookDAO {
 	public List<Book> getAllBooks();
-	public Book getBookById(int bookId);
-	public List<Book> getBooksByUser(int userId);
+	public Book getBookById(long bookId);
+	public List<Book> getBooksByUser(long userId);
 	public List<Book> getBooksByUserNull();
 	public void addBook(Book book);
 	public boolean modifyBook(Book book);
