+++ src/main/java/com/nashtech/intern/dao/IUserDAO.java
@@ -6,7 +6,7 @@ import com.nashtech.intern.model.User;
 
 public interface IUserDAO {
 	public List<User> getAllUser();
-	public User getUserById(int userId);
+	public User getUserById(long userId);
 	public void addUser(User user);
 	public boolean modifyUser(User user);
 }
